requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ollama>=0.1.0
langchain>=0.0.300
pillow>=10.0.0
watchdog>=3.0.0
openai>=1.0.0
llama-index>=0.9.0
pyyaml>=6.0.0